+++ globals.py
@@ -1,3 +1,1 @@
-def initialize():
-    global active_users
-    active_users = {}
+active_users = {}