+++ utils/__init__.py
@@ -1,3 +1,2 @@
 from .getRegex import get_regex
 from .generateString import generate_string
-from .pushTele import push_tele