+++ versions.example.py
@@ -1,0 +1,7 @@
+versionAvailable = [
+   'your',
+   'allowed',
+   'app',
+   'version',
+   'here'
+]