+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## Version 0.2.0
+- Add feature creating HubSpot contact via Telegram.
+- Fix not logging AhaMove's users profile which slowed down scraping ads process.
+- Fix not expanding content of posts with background.
+- Speed up scraping Ads process.
+- Little changes in database.
+- Adds API to have more utilities for staff: Check contact's status, Create HubSpot contact via Telegram.
+
 ## Version 0.1.0
 - Beta version